Move coinListOf method to BitcoinTestUtils
